+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 * Explicitly forbid datetime interval conditions (#373).
+* Storage initialization is now asynchronous by default for Tarantool 3.0+ (#412).
 
 ### Fixed
 * Working with datetime conditions in case of non-indexed fields or